Rename camel case files
