+++ app/components/layout.js
@@ -1,6 +1,6 @@
 import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
-import Navbar from '../components/common/Navbar';
+import Navbar from '../components/common/navbar';
 import {connectToServer, disconnectFromServer, handshake, startCpuProfiling, stopCpuProfiling} from '../redux/actions';
 import DisplayHandshake from './display_handshake';
 import Toolbelt from './toolbelt';
