+++ app/rbkit_connection.js
